requirements.txt
tiktoken==0.7.0
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
PyMySQL
crewai[google-genai]
pandas
scikit-learn
python-dotenv
cryptography
google-generativeai
firebase-admin
gunicorn
fpdf2litellm
